Rename 'exporter' to 'disclosure'
