+++ lca_disclosures/base/__init__.py
@@ -1,0 +1,1 @@
+from .disclosure import BaseDisclosure
