+++ lca_disclosures/brightway2/disclosure.py
@@ -1,7 +1,7 @@
 import brightway2 as bw
 import numpy as np
 
-from ..base.exporter import BaseDisclosure
+from ..base import BaseDisclosure
 from ..utils import matrix_to_coo
